Fix #65 by introducing vuex
