+++ packages/server-web/package.json
@@ -50,6 +50,7 @@
     "passport-local": "^1.0.0",
     "vue": "^2.5.16",
     "vue-resource": "1.5.1",
-    "vue-router": "^3.0.1"
+    "vue-router": "^3.0.1",
+    "vuex": "^3.0.1"
   }
 }
