+++ packages/server-web/src/client/client.js
@@ -1,6 +1,6 @@
 import createSocketConnection from "./socket-connection.js";
 import createApp from "./create-app.js";
-import { processEvent } from "./store.js";
+import { processEvent } from "./todo-store.js";
 
 const commands = createSocketConnection(processEvent);
 createApp(commands);
