+++ packages/server-web/src/client/create-app.js
@@ -4,6 +4,7 @@ import LoginComponent from "./component/Login.vue";
 import HomeComponent from "./component/Home.vue";
 import AppComponent from "./component/App.vue";
 import VueResource from "vue-resource";
+import { store } from "./store.js";
 
 Vue.use(Router);
 Vue.use(VueResource);
@@ -37,6 +38,7 @@ export default function createApp(commands) {
   const router = createRouter(commands);
   new Vue({
     el: "#app",
+    store,
     render: h => h(AppComponent, { props: { commands } }),
     router
   });
